+++ projects/core/src/sql-command.ts
@@ -1,14 +1,17 @@
 import type { EntityMetadata } from './remult3/remult3.js'
 
-export interface SqlImplementation {
+export interface SqlImplementation extends HasWrapIdentifier {
   getLimitSqlSyntax(limit: number, offset: number)
   createCommand(): SqlCommand
   transaction(action: (sql: SqlImplementation) => Promise<void>): Promise<void>
   entityIsUsedForTheFirstTime(entity: EntityMetadata): Promise<void>
   ensureSchema?(entities: EntityMetadata[]): Promise<void>
   supportsJsonColumnType?: boolean
-  wrapIdentifier?(name: string): string
+
   afterMutation?: VoidFunction
+}
+export interface HasWrapIdentifier {
+  wrapIdentifier?(name: string): string
 }
 
 export interface SqlCommand extends SqlCommandWithParameters {