+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.25.6] TBD
+
+- improved dbNamesOf of to use by default the wrapIdentifier of the current data provider if no wrap identifier was provided
+- Added support for using dbNamesOf in an sql expression for that same entity
+
 ## [0.25.5] 2024-02-11
 
 - Added `admin` option to servers, enabling the `/api/admin` route with a built in entity explorer